Improve tests (#8)
* Improve tests.
* Fix `transforms.functional.normalize` dtype cast.
* Fix several `__repr__` for transforms.
* Remove None support for `RandomColorJitter` and `BatchRandomColorJitter`

+++ torchaug/transforms/functional.py
@@ -275,7 +275,7 @@ def normalize(
         if not torch.tensor(0, dtype=cast_dtype).is_floating_point():
             raise ValueError(f"cast_dtype should be a float dtype. Got {cast_dtype}.")
         casted = True
-        tensor = convert_image_dtype(tensor, dtype=torch.float32)
+        tensor = convert_image_dtype(tensor, dtype=cast_dtype)
     else:
         casted = False
